+++ pyfusa/__init__.py
@@ -11,7 +11,7 @@ import unicodedata
 from dataclasses import dataclass, field
 from typing import Optional
 
-VERSION = "0.1.9"
+VERSION = "0.2.0"
 SPEC_VERSION = "1.10.4"
 LANGUAGE = "python"
 TOOL = "py-FuSa"